# src/main/openapi/recipe-api-v1.yaml
openapi: 3.0.1
info:
  title: Recipe Api
  description: Recipe Management Api's
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /admin/user:
    post:
      tags:
        - user-controller
      summary: 'Register the new user; Role Required: ROLE_MANAGER'
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '201':
          description: Return the created User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          description: Internal Server Exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        subErrors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorResponse'
    FieldError:
      type: object
      properties:
        codes:
          type: array
          items:
            type: string
        arguments:
          type: array
          items:
            type: object
        defaultMessage:
          type: string
        objectName:
          type: string
        field:
          type: string
        rejectedValue:
          type: object
        bindingFailure:
          type: boolean
        code:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        password:
          type: string
        userRoles:
          uniqueItems: true
          type: array
          items:
            type: string